% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesclassifier.R
\name{bayesclassifier}
\alias{bayesclassifier}
\title{Bayes' rule of allocation}
\usage{
bayesclassifier(dat, params)
}
\arguments{
\item{dat}{An \eqn{n\times p} matrix where each row represents an individual observation}

\item{params}{A list containing the required parameters (p, g, pi, mu, and sigma) from sslparlist() or provided manually.}
}
\value{
\item{clust}{Class membership for the ith entity}
}
\description{
Bayes' rule of allocation
}
\details{
Classifier specified by Bayes' rule


The classifier/Bayes rule of allocation \eqn{R(y_j;\theta)} assigns an entity with observation \eqn{y_j} to  class \eqn{C_k }(that is, \eqn{R(y_j;\theta)=k}) if
\eqn{	k=\arg\max_i \tau_i(y_j;\theta),}
}
\examples{
n <- 150
pi <- c(0.25, 0.25, 0.25, 0.25)
sigma <- array(0, dim = c(3, 3, 4))
sigma[, , 1] <- diag(1, 3)
sigma[, , 2] <- diag(2, 3)
sigma[, , 3] <- diag(3, 3)
sigma[, , 4] <- diag(4, 3)
mu <- matrix(c(0.2, 0.3, 0.4, 0.2, 0.7, 0.6, 0.1, 0.7, 1.6, 0.2, 1.7, 0.6), 3, 4)
dat <- rmix(n = n, pi = pi, mu = mu, sigma = sigma)
params <- list(p = 3, g = 4, pi = pi, mu = mu, sigma = sigma)
clust <- bayesclassifier(dat = dat$Y, params = params)
}
