% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov2vec.R
\name{cov2vec}
\alias{cov2vec}
\title{Transform a variance matrix into a vector}
\usage{
cov2vec(sigma)
}
\arguments{
\item{sigma}{A \eqn{p\times p} variance matrix}
}
\value{
par A vector representing a variance matrix
}
\description{
Transform a variance matrix into a vector i.e., Sigma=R^T*R
}
\details{
The variance matrix is decomposed by computing the Choleski factorization of a real symmetric positive-definite square matrix.
Then, storing the upper triangular factor of the Choleski decomposition into a vector.
}
