% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/erate.R
\name{erate}
\alias{erate}
\title{Error rate of the Bayes rule for a g-class Gaussian mixture model}
\usage{
erate(dat, parlist, clust)
}
\arguments{
\item{dat}{An \eqn{n\times p} matrix where each row represents an individual observation}

\item{parlist}{A list containing the required parameters (p, g, pi, mu, and sigma) from sslparlist() or provided manually.}

\item{clust}{An n-dimensional vector of class partition.}
}
\value{
\item{errval}{a value of error rate}
}
\description{
Error rate of the Bayes rule for a g-class Gaussian mixture model
}
\details{
The  error rate of the Bayes rule for a g-class Gaussian mixture model is given by
\deqn{
err(y;\theta)=1-\sum_{i=1}^g\pi_i Pr\{R(y;\theta)=i\mid Z \in C_i\}.
}
Here, we write
\deqn{
Pr\{R(y;\theta) \in C_i\mid Z\in C_i\}=\frac{\sum_{j=1}^nI_{C_i}(z_j)Q[z_j,R(y;\theta) ]}{\sum_{j=1}^nI_{C_i}(z_j)},
}
where \eqn{Q[u,v]=1} if \eqn{u=v} and \eqn{Q[u,v]=0} otherwise, and \eqn{I_{C_i}(z_j)} is an indicator function for the \eqn{i}th class.
}
\examples{
n<-150
pi<-c(0.25,0.25,0.25,0.25)
sigma<-array(0,dim=c(3,3,4))
sigma[,,1]<-diag(1,3)
sigma[,,2]<-diag(2,3)
sigma[,,3]<-diag(3,3)
sigma[,,4]<-diag(4,3)
mu<-matrix(c(0.2,0.3,0.4,0.2,0.7,0.6,0.1,0.7,1.6,0.2,1.7,0.6),3,4)
dat<-rmix(n=n,pi=pi,mu=mu,sigma=sigma)
xi<-c(-0.5,1)
m<-rlabel(dat=dat$Y,pi=pi,mu=mu,sigma=sigma,xi=xi)
zm<-dat$clust
zm[m==1]<-NA
inits<-initialvalue(g=4,zm=zm,dat=dat$Y)
\donttest{
fit_pc<-gmmsslm(dat=dat$Y,zm=zm,pi=inits$pi,mu=inits$mu,sigma=inits$sigma,xi=xi,type='full')
parlist<-sslparlist(fit_pc)
erate(dat=dat$Y,parlist=parlist,clust=dat$clust)
}

}
