% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gastro_data.R
\docType{data}
\name{gastro_data}
\alias{gastro_data}
\title{Gastrointestinal dataset}
\description{
The collected dataset is composed of 76 colonoscopic videos (recorded with both White Light (WL) and Narrow Band Imaging (NBI)), the histology (classification ground truth), and  the endoscopist's opinion (including 4 experts and 3 beginners). There are $n=76$ observations, and each observation consists of 698 features extracted from colonoscopic videos on patients with gastrointestinal lesions.
}
\references{
\url{http://www.depeca.uah.es/colonoscopy_dataset/}
}
