% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmsslm.R
\name{gmmsslm}
\alias{gmmsslm}
\title{Fitting Gaussian mixture model to a complete classified dataset or an incomplete classified dataset with/without the missing-data mechanism.}
\usage{
gmmsslm(
  dat,
  zm,
  pi,
  mu,
  sigma,
  xi = NULL,
  type,
  iter.max = 500,
  eval.max = 500,
  rel.tol = 1e-15,
  sing.tol = 1e-15
)
}
\arguments{
\item{dat}{An \eqn{n\times p} matrix where each row represents an individual observation}

\item{zm}{An n-dimensional vector containing the class labels including the missing-label denoted as NA.}

\item{pi}{A g-dimensional vector for the initial values of the mixing proportions.}

\item{mu}{A \eqn{p \times g} matrix for the initial values of the location parameters.}

\item{sigma}{A \eqn{p\times p} covariance matrix,or a list of g covariance matrices with dimension \eqn{p\times p \times g}.
It is assumed to fit the model with a common covariance matrix if \code{sigma} is a \eqn{p\times p} covariance matrix;
otherwise it is assumed to fit the model with unequal covariance matrices.}

\item{xi}{A 2-dimensional vector containing the initial values of the coefficients in the logistic function of the Shannon entropy.}

\item{type}{Three types of Gaussian mixture models, 'ign' indicates fitting the model to a partially classified sample on the basis of the likelihood that ignores the missing label mechanism,
'full' indicates fitting the model to a partially classified sample on the basis of the full likelihood, taking into account the missing-label mechanism,
and 'com' indicate fitting the model to a completed classified sample.}

\item{iter.max}{Maximum number of iterations allowed. Defaults to 500}

\item{eval.max}{Maximum number of evaluations of the objective function allowed. Defaults to 500}

\item{rel.tol}{Relative tolerance. Defaults to 1e-15}

\item{sing.tol}{Singular convergence tolerance; defaults to 1e-20.}
}
\value{
A gmmsslmFit object containing the following slots:
\item{objective}{Value of objective likelihood}
\item{convergence}{Value of convergence}
\item{iteration}{Number of iterations}
\item{obs}{Input data matrix}
\item{n}{Number of observations}
\item{p}{Number of variables}
\item{g}{Number of Gaussian components}
\item{type}{Type of Gaussian mixture model}
\item{pi}{Estimated vector of the mixing proportions}
\item{mu}{Estimated matrix of the location parameters}
\item{sigma}{Estimated covariance matrix or list of covariance matrices}
\item{xi}{Estimated coefficient vector for a logistic function of the Shannon entropy}
}
\description{
Fitting Gaussian mixture model to a complete classified dataset or an incomplete classified dataset with/without the missing-data mechanism.
}
\examples{
n<-150
pi<-c(0.25,0.25,0.25,0.25)
sigma<-array(0,dim=c(3,3,4))
sigma[,,1]<-diag(1,3)
sigma[,,2]<-diag(2,3)
sigma[,,3]<-diag(3,3)
sigma[,,4]<-diag(4,3)
mu<-matrix(c(0.2,0.3,0.4,0.2,0.7,0.6,0.1,0.7,1.6,0.2,1.7,0.6),3,4)
dat<-rmix(n=n,pi=pi,mu=mu,sigma=sigma)
xi<-c(-0.5,1)
m<-rlabel(dat=dat$Y,pi=pi,mu=mu,sigma=sigma,xi=xi)
zm<-dat$clust
zm[m==1]<-NA
inits<-initialvalue(g=4,zm=zm,dat=dat$Y)
\donttest{
fit_pc<-gmmsslm(dat=dat$Y,zm=zm,pi=inits$pi,mu=inits$mu,sigma=inits$sigma,xi=xi,type='full')
}
}
