% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmsslm.R
\docType{class}
\name{gmmsslmFit-class}
\alias{gmmsslmFit-class}
\title{gmmsslmFit Class}
\description{
gmmsslmFit objects store the results of fitting Gaussian mixture models using the gmmsslm function.

An S4 class representing the result of fitting a Gaussian mixture model using gmmsslm()
}
\section{Slots}{

\describe{
\item{\code{objective}}{A numeric value representing the objective likelihood.}

\item{\code{ncov}}{A numeric value representing the number of covariance matrices.}

\item{\code{convergence}}{A numeric value representing the convergence value.}

\item{\code{iteration}}{An integer value representing the number of iterations.}

\item{\code{obs}}{A matrix containing the input data.}

\item{\code{n}}{An integer value representing the number of observations.}

\item{\code{p}}{An integer value representing the number of variables.}

\item{\code{g}}{An integer value representing the number of Gaussian components.}

\item{\code{type}}{A character value representing the type of Gaussian mixture model.}

\item{\code{pi}}{A numeric vector representing the mixing proportions.}

\item{\code{mu}}{A matrix representing the location parameters.}

\item{\code{sigma}}{An array representing the covariance matrix or list of covariance matrices.}

\item{\code{xi}}{A numeric value representing the coefficient for a logistic function of the Shannon entropy.}
}}

\seealso{
gmmsslm
}
