% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logsumexp.R
\name{logsumexp}
\alias{logsumexp}
\title{log summation of exponential function}
\usage{
logsumexp(x)
}
\arguments{
\item{x}{A variable vector.}
}
\value{
\item{val}{log summation of exponential variable vector.}
}
\description{
log summation of exponential variable vector.
}
