% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalise_logprob.R
\name{normalise_logprob}
\alias{normalise_logprob}
\title{Normalize log-probability}
\usage{
normalise_logprob(x)
}
\arguments{
\item{x}{A variable vector.}
}
\value{
\item{val}{A normalize log probability of variable vector.}
}
\description{
Normalize log-probability.
}
