% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmsslm.R
\name{predict}
\alias{predict}
\alias{predict,gmmsslmFit-method}
\title{Predict unclassified label}
\usage{
predict(object)
}
\arguments{
\item{object}{A gmmsslmFit object.}
}
\description{
This function predicts unclassified label from a gmmsslmFit object.
}
