% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pro2vec.R
\name{pro2vec}
\alias{pro2vec}
\title{Transfer a probability vector into a vector}
\usage{
pro2vec(pro)
}
\arguments{
\item{pro}{An propability vector}
}
\value{
y An informative vector
}
\description{
Transfer a probability vector into an informative vector
}
