% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmsslm.R
\name{sslparlist}
\alias{sslparlist}
\alias{sslparlist,gmmsslmFit-method}
\title{Extract parameter list from gmmsslmFit objects}
\usage{
sslparlist(object)
}
\arguments{
\item{object}{A gmmsslmFit object.}
}
\description{
This function extracts the parameters from a gmmsslmFit object, including p, g, pi, mu, and sigma.
}
