% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gmmsslm.R
\name{summary}
\alias{summary}
\alias{summary,gmmsslmFit-method}
\title{Summary method for gmmsslmFit objects}
\usage{
summary(object)
}
\arguments{
\item{object}{A gmmsslmFit object.}
}
\description{
This function extracts summary information from a gmmsslmFit object, including objective value, ncov, convergence, iteration, and type.
}
