% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2cov.R
\name{vec2cov}
\alias{vec2cov}
\title{Transform a vector into a matrix}
\usage{
vec2cov(par)
}
\arguments{
\item{par}{A vector representing a variance matrix}
}
\value{
sigma A variance matrix
}
\description{
Transform a vector into a matrix i.e., Sigma=R^T*R
}
\details{
The variance matrix is decomposed by computing the Choleski factorization of a real symmetric positive-definite square matrix.
Then, storing the upper triangular factor of the Choleski decomposition into a vector.
}
