% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vec2pro.R
\name{vec2pro}
\alias{vec2pro}
\title{Transfer an informative vector to a probability vector}
\usage{
vec2pro(vec)
}
\arguments{
\item{vec}{An informative vector}
}
\value{
pro A probability vector
}
\description{
Transfer an informative vector to a probability vector
}
